FX cleanup: move all *Def classes to fx_defs.h
