+++ fx_manager.cpp
@@ -1,9 +1,7 @@
 #include "fx_manager.h"
 
 #include "fx_color.h"
-#include "fx_emitter_def.h"
 #include "fx_math.h"
-#include "fx_particle_def.h"
 #include "fx_particle_system.h"
 #include "fx_rect.h"
