+++ fx_simple.h
@@ -2,7 +2,7 @@
 
 #include "util.h"
 
-class Color;
+struct Color;
 
 RICH_ENUM(FXName, TEST_SIMPLE, TEST_MULTI, WOOD_SPLINTERS, ROCK_SPLINTERS, ROCK_CLOUD, EXPLSION, RIPPLE, FEET_DUST,
           FIRE, CIRCULAR_BLAST, MAGIC_MISSILE, FIREBALL, SLEEP, INSANITY, BLIND, PEACEFULNESS, SLOW, SPEED, FLYING);
